+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.14.2
 
 * Adds support for `onMapLoaded` callback.
+* Adds `onMapLoaded` and `onGroundOverlayTap` implementations to `MethodChannelGoogleMapsFlutter`.
 
 ## 2.14.1